scripts/requirements.txt
pillow-heif
pillow
exifread
pyheif